+++ gnl/get_next_line_utils_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line_utils_bonus.c                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sgi <sgi@student.42seoul.kr>               +#+  +:+       +#+        */
+/*   By: sgi <sgi@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 10:46:44 by sgi               #+#    #+#             */
-/*   Updated: 2022/07/01 10:46:45 by sgi              ###   ########.fr       */
+/*   Updated: 2022/08/01 21:13:18 by sgi              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ int	gnl_read(int fd, t_status *s)
 
 	bytes = read(fd, s->buf, BUFFER_SIZE);
 	s->read_bytes = 0;
-	if (bytes < BUFFER_SIZE)
+	if ((unsigned int)bytes < BUFFER_SIZE)
 		s->buf[bytes] = EOF;
 	return (bytes);
 }
